created utils module as part of the package
